Move test config into test_common
Make it easier to import from test_manual.

Also fix an incorrect type annotation
